+++ js/render.js
@@ -1,0 +1,12 @@
+/**
+ * Создает DOM-элемент по шаблону.
+ * @param {string} markup Разметка элемента.
+ * @return {Node} Новый элемент.
+ */
+const render = (markup) => {
+  const template = document.createElement(`template`);
+  template.innerHTML = markup.trim();
+  return template.content.firstChild;
+};
+
+export default render;
