+++ js/util.js
@@ -1,5 +1,3 @@
-const app = document.querySelector(`div.app`);
-
 // Варианты склонения для 1, 2-4, 0,5+
 const cases = {
   minutes: [`минуту`, `минуты`, `минут`],
@@ -8,26 +6,6 @@ const cases = {
   mistakes: [`ошибку`, `ошибки`, `ошибок`],
   attempts: [`раз`, `раза`, `раз`],
   fast: [`быстрый`, `быстрых`, `быстрых`]
-};
-
-/**
- * Создает DOM-элемент по шаблону.
- * @param {string} markup Разметка элемента.
- * @return {Node} Новый элемент.
- */
-export const render = (markup) => {
-  const template = document.createElement(`template`);
-  template.innerHTML = markup.trim();
-  return template.content.firstChild;
-};
-
-/**
- * Переключает текущий экран.
- * @param {Node} screen Экран, который нужно отобразить.
- */
-export const changeScreen = (screen) => {
-  const currentScreen = app.querySelector(`section.main`);
-  app.replaceChild(screen, currentScreen);
 };
 
 /**
